Modified abstract, added names.
Added our names to main.tex. Please check for Yuri's and Marcelo's institutions.

Largely changed the structure of the abstract. The focus of the paper should be
on the analysis of the contribution of each of the three adaptations (representation,
hybridization, clustering) to the generation of risk models.

Please take this into account while re-writing the text.